FIX: refactoring code, split into different modules, deduplicating some of the wasm stuff as it exists elsewhere, adding sp1-wasm to the project for the gnark conversion methods. TODO: need to implement O1jsProof and O1jsVK from_arkworks methods so we can load them into sp1_verifier.
