+++ pairing-utils/src/lib.rs
@@ -9,7 +9,9 @@ pub mod eth_root;
 pub mod kzg;
 mod risc0_vk;
 pub mod serialize;
+pub mod snarkjs;
 pub mod tonelli_shanks;
+pub mod types;
 pub mod utils;
 pub mod write;
 
@@ -17,6 +19,11 @@ pub mod write;
 pub mod wasm;
 //#[cfg(feature = "wasm")]
 pub use wasm::{compute_and_serialize_aux_witness_js, compute_pairing_js};
+
+// Re-export commonly used types
+pub use serialize::{AuxWitness, Field12};
+pub use snarkjs::{O1jsProof, O1jsVK, SnarkjsProof, SnarkjsVK};
+pub use types::{AffinePoint2d, ComplexAffinePoint2d, ComplexProjectivePoint, ProjectivePoint};
 
 pub fn display_fq12(x: Fq12, label: &str) {
     println!("{}.g00: {}", label, x.c0.c0.c0);
